Move to fixed point

+++ src/core/game_types.h
@@ -30,19 +30,19 @@ typedef struct {
 struct Player {
     int x;  // Fixed-point
     int y;  // Fixed-point
-    float vx; // Floating-point
-    float vy; // Floating-point
-    float maxFall; // Current max fall speed (ramps for fast fall)
+    int vx; // Fixed-point
+    int vy; // Fixed-point
+    int maxFall; // Current max fall speed (ramps for fast fall)
     int onGround;
     int wasOnGround;  // Previous frame onGround (for lift boost detection)
     int coyoteTime;   // Frames remaining for coyote time jump
     int jumpBuffer;   // Frames remaining in jump buffer
     int jumpHeld;     // 1 if jump was initiated this aerial phase
     int autoJump;     // 1 if AutoJump is active (simulates holding jump button for maintained jump height)
-    float autoJumpTimer;  // Frames remaining for AutoJump window
-    float liftBoostX; // Velocity from moving platforms (X component)
-    float liftBoostY; // Velocity from moving platforms (Y component)
-    float varJumpSpeed; // Initial jump velocity for var jump clamping
+    int autoJumpTimer;  // Frames remaining for AutoJump window
+    int liftBoostX; // Velocity from moving platforms (X component)
+    int liftBoostY; // Velocity from moving platforms (Y component)
+    int varJumpSpeed; // Initial jump velocity for var jump clamping
     int varJumpTimer;   // Frames remaining for variable jump window
     int dashing;
     int dashes;                // Current dashes available (0-2)
@@ -53,18 +53,18 @@ struct Player {
     u16 prevKeys;     // Previous frame keys for detecting button presses
 
     // Wall slide/jump
-    float wallSlideTimer;  // Time remaining for wall slide ability (resets on ground/jump)
+    int wallSlideTimer;  // Time remaining for wall slide ability (resets on ground/jump)
     int wallSlideDir;      // Direction of current wall slide: 1=right, -1=left, 0=none
 
     // Dash attack (for super jumps)
     int dashAttackTimer;   // Frames remaining where super jumps can be triggered (set on dash start)
     int dashDirX;          // Dash direction X: -1=left, 0=none, 1=right
     int dashDirY;          // Dash direction Y: -1=up, 0=none, 1=down
-    float beforeDashSpeedX;  // Horizontal speed before dash (for preserving higher speeds)
+    int beforeDashSpeedX;  // Horizontal speed before dash (for preserving higher speeds)
     int ducking;           // 1 if player is ducking (affects super jump height)
 
     // Climbing
-    float stamina;         // Current stamina (0-110)
+    int stamina;         // Current stamina (0-28160 in fixed-point, 0-110 logical)
     int climbNoMoveTimer;  // Frames before climb movement is allowed (prevents grab spam)
     int lastClimbMove;     // Last climb movement direction: -1=up, 0=still, 1=down
     int wallBoostTimer;    // Frames remaining to convert climb jump to wall jump (refunds stamina)
@@ -72,9 +72,9 @@ struct Player {
 
     // Climb hop system (Celeste Player.cs line 218-219)
     int hopWaitX;          // If you climb hop onto a solid, snap beside it until you get above it
-    float hopWaitXSpeed;   // Horizontal speed to apply when hopWaitX releases
+    int hopWaitXSpeed;   // Horizontal speed to apply when hopWaitX releases
     int forceMoveX;        // Force horizontal input to this value
-    float forceMoveXTimer; // Frames remaining to force horizontal input
+    int forceMoveXTimer; // Frames remaining to force horizontal input
 
     // Dash trail
     int trailX[TRAIL_LENGTH];  // Fixed-point positions
